Game Feel Improvement
Player now no longer travels through walls with movement mechanics. There is a basic version of the battery implemented to add challenge to the game. The UI has a message that pops up when the drones attack. The drones move around a bit rather than staying static (drones = those cubes).