Menus and Bug Fixes
Fixed a few bugs with the Scanning and something to do with the Gatling Guns on the salamander. Implemented a Main Menu with working buttons, implemented a pause menu, a little buggy, will need to revise, it just uses Time.timeScale = 0, should rework all other scripts to feed through GameManager so that pause will stop script. Cause of this the pause menu buttons don't work, so when you pause ('p'), then hit 'q' to Quit.