Basic QOL Improvements
I fixed the scanning distance bug, I added Victory and Defeat text to the win/lose condition so the player knows (and the dead model is destroyed). I wasn't able to fix the performance issue but I had a look at the code of Sal and it doesn't seem like it should put a strain on the game, it is strange, seems to happen less frequently though.