Dragon fly FPS issue fix + salamander increase + replenish object
Fixed the dragon fly FPS issue - the dragon fly now removes the collider in the event that it intersects with the salamander and reapplies the collider upon exit of intersection. It periodically checks around itself to avoid the player/salamander/walls. Everything else remains unchanged.

Salamander is 2x size with pathways adjusted so they don't look wack with it's new size.

Piles of rubbish introduced as objects which constantly check for player within a radius. If spotted they move towards the player and upon contact replenish shield and delete themselves.