Laser Implementation / Robot Prefab Cleanup
Added laser objects attached to the robot finger.
cleaned up the robot prefab to exclude a few objects not needed and fixed the right hand finger issue.
Setup of the laser extension added. (need to actually add the extension now).
Fixed Jon's smoke and added them to the right position in the hierarchy.